3619: Rename EntityAttributesVariableStore files
