+++ common/src/Model/EntityPropertiesVariableStore.h
